feat(core): extract config/ into core/ kernel (Phase A)
Move config/index.ts, config/env.schema.ts, and 3 colocated test files
to backend/src/core/config/. Create re-export shims at original paths
so all existing consumers continue to resolve without changes.

Part of #706
