+++ backend/src/core/config/env.schema.ts
@@ -1,0 +1,241 @@
+import { z } from 'zod';
+
+/** Optional URL that treats empty strings as undefined (common in Docker Compose env defaults). */
+const optionalUrl = z.preprocess(
+  (v) => (typeof v === 'string' && v.trim() === '' ? undefined : v),
+  z.string().url().optional(),
+);
+
+export const envSchema = z.object({
+  // Auth
+  DASHBOARD_USERNAME: z.string().min(1),
+  DASHBOARD_PASSWORD: z.string().min(12),
+  JWT_SECRET: z.string().min(32),
+  // JWT signing algorithm: HS256 (symmetric, default), RS256 or ES256 (asymmetric).
+  // HS256 is appropriate for single-service architectures. Switch to RS256/ES256 if:
+  //   - Multiple backend services need to verify tokens independently
+  //   - Token verification moves to edge/proxy layers
+  //   - Compliance mandates asymmetric signing
+  // When using RS256/ES256, JWT_SECRET is ignored and JWT_PRIVATE_KEY_PATH +
+  // JWT_PUBLIC_KEY_PATH must be set to PEM key files.
+  JWT_ALGORITHM: z.enum(['HS256', 'RS256', 'ES256']).default('HS256'),
+  JWT_PRIVATE_KEY_PATH: z.string().optional(),
+  JWT_PUBLIC_KEY_PATH: z.string().optional(),
+
+  // Portainer
+  PORTAINER_API_URL: z.string().url().default('http://localhost:9000'),
+  PORTAINER_API_KEY: z.string().default(''),
+  PORTAINER_VERIFY_SSL: z.string().default('true').transform((v) => v === 'true' || v === '1'),
+  PORTAINER_CONCURRENCY: z.coerce.number().int().min(1).max(100).default(30),
+  PORTAINER_MAX_CONNECTIONS: z.coerce.number().int().min(1).max(100).default(20),
+  PORTAINER_CB_FAILURE_THRESHOLD: z.coerce.number().int().min(1).max(50).default(5),
+  PORTAINER_CB_RESET_TIMEOUT_MS: z.coerce.number().int().min(1000).max(300000).default(30000),
+  // Public/external dashboard URL used by remote agents/endpoints to call back into the API.
+  DASHBOARD_EXTERNAL_URL: optionalUrl,
+
+  // Ollama / LLM
+  OLLAMA_BASE_URL: z.string().url().default('http://host.docker.internal:11434'),
+  OLLAMA_MODEL: z.string().default('llama3.2'),
+  LLM_OPENAI_ENDPOINT: optionalUrl, // OpenAI-compatible endpoint (e.g., OpenWebUI)
+  LLM_BEARER_TOKEN: z.string().optional(), // Bearer token or username:password for Basic auth
+  LLM_AUTH_TYPE: z.enum(['bearer', 'basic']).default('bearer'), // Auth header type for LLM endpoint tokens
+  LLM_VERIFY_SSL: z.string().default('true').transform((v) => v === 'true' || v === '1'),
+  LLM_REQUEST_TIMEOUT: z.coerce.number().int().min(5000).max(600000).default(120000),
+
+  // AI Search specific model (defaults to llama3.2:latest - change to qwen3:30b for better reasoning)
+  AI_SEARCH_MODEL: z.string().default('llama3.2:latest'),
+
+  // Kibana (optional)
+  KIBANA_ENDPOINT: z.string().url().optional(),
+  KIBANA_API_KEY: z.string().optional(),
+
+  // Monitoring
+  MONITORING_ENABLED: z.coerce.boolean().default(true),
+  MONITORING_INTERVAL_MINUTES: z.coerce.number().int().min(1).default(5),
+  MAX_INSIGHTS_PER_CYCLE: z.coerce.number().int().min(1).max(10000).default(500),
+  AI_ANALYSIS_ENABLED: z.coerce.boolean().default(true),
+
+  // Metrics Collection
+  METRICS_COLLECTION_ENABLED: z.coerce.boolean().default(true),
+  METRICS_COLLECTION_INTERVAL_SECONDS: z.coerce.number().int().min(10).default(60),
+  METRICS_RETENTION_DAYS: z.coerce.number().int().min(1).default(7),
+  METRICS_ENDPOINT_CONCURRENCY: z.coerce.number().int().min(1).max(50).default(10),
+  METRICS_CONTAINER_CONCURRENCY: z.coerce.number().int().min(1).max(100).default(20),
+  PROMETHEUS_METRICS_ENABLED: z.coerce.boolean().default(false),
+  PROMETHEUS_BEARER_TOKEN: z.string().optional(),
+
+  // Anomaly Detection
+  ANOMALY_ZSCORE_THRESHOLD: z.coerce.number().min(0.5).default(3.5),
+  ANOMALY_MOVING_AVERAGE_WINDOW: z.coerce.number().int().min(5).default(20),
+  ANOMALY_MIN_SAMPLES: z.coerce.number().int().min(3).default(10),
+  ANOMALY_DETECTION_METHOD: z.enum(['zscore', 'bollinger', 'adaptive']).default('adaptive'),
+  ANOMALY_COOLDOWN_MINUTES: z.coerce.number().int().min(0).default(30),
+  ANOMALY_THRESHOLD_PCT: z.coerce.number().min(50).max(100).default(85),
+  ANOMALY_HARD_THRESHOLD_ENABLED: z.coerce.boolean().default(true),
+  BOLLINGER_BANDS_ENABLED: z.coerce.boolean().default(true),
+
+  // Predictive Alerting
+  PREDICTIVE_ALERTING_ENABLED: z.coerce.boolean().default(true),
+  PREDICTIVE_ALERT_THRESHOLD_HOURS: z.coerce.number().int().min(1).default(24),
+
+  // Anomaly Explanations (LLM)
+  ANOMALY_EXPLANATION_ENABLED: z.coerce.boolean().default(true),
+  ANOMALY_EXPLANATION_MAX_PER_CYCLE: z.coerce.number().int().min(1).max(50).default(5),
+
+  // Isolation Forest Anomaly Detection
+  ISOLATION_FOREST_ENABLED: z.coerce.boolean().default(true),
+  ISOLATION_FOREST_TREES: z.coerce.number().int().min(10).max(500).default(100),
+  ISOLATION_FOREST_SAMPLE_SIZE: z.coerce.number().int().min(32).max(512).default(256),
+  ISOLATION_FOREST_CONTAMINATION: z.coerce.number().min(0.01).max(0.5).default(0.15),
+  ISOLATION_FOREST_RETRAIN_HOURS: z.coerce.number().int().min(1).default(6),
+
+  // NLP Log Analysis (LLM)
+  NLP_LOG_ANALYSIS_ENABLED: z.coerce.boolean().default(true),
+  NLP_LOG_ANALYSIS_MAX_PER_CYCLE: z.coerce.number().int().min(1).max(20).default(3),
+  NLP_LOG_ANALYSIS_TAIL_LINES: z.coerce.number().int().min(10).max(500).default(100),
+
+  // Smart Alert Grouping
+  SMART_GROUPING_ENABLED: z.coerce.boolean().default(true),
+  SMART_GROUPING_SIMILARITY_THRESHOLD: z.coerce.number().min(0.1).max(1.0).default(0.3),
+  INCIDENT_SUMMARY_ENABLED: z.coerce.boolean().default(true),
+
+  // Investigation (Root Cause Analysis)
+  INVESTIGATION_ENABLED: z.coerce.boolean().default(true),
+  INVESTIGATION_COOLDOWN_MINUTES: z.coerce.number().int().min(1).default(20),
+  INVESTIGATION_MAX_CONCURRENT: z.coerce.number().int().min(1).default(2),
+  INVESTIGATION_LOG_TAIL_LINES: z.coerce.number().int().min(10).default(50),
+  INVESTIGATION_METRICS_WINDOW_MINUTES: z.coerce.number().int().min(5).default(60),
+  INVESTIGATION_MIN_SEVERITY: z.enum(['critical', 'warning', 'info']).default('warning'),
+
+  // Packet Capture (PCAP)
+  PCAP_ENABLED: z.coerce.boolean().default(false),
+  PCAP_MAX_DURATION_SECONDS: z.coerce.number().int().min(1).max(3600).default(300),
+  PCAP_MAX_FILE_SIZE_MB: z.coerce.number().int().min(1).max(500).default(50),
+  PCAP_MAX_CONCURRENT: z.coerce.number().int().min(1).max(10).default(2),
+  PCAP_RETENTION_DAYS: z.coerce.number().int().min(1).default(7),
+  PCAP_STORAGE_DIR: z.string().default('./data/pcap'),
+  PCAP_CAPTURE_IMAGE: z.string().default('alpine:3.21'),
+  PCAP_CAPTURE_IMAGE_PULL: z.enum(['always', 'never', 'if-not-present']).default('if-not-present'),
+
+  // Remediation Safety
+  REMEDIATION_PROTECTED_CONTAINERS: z.string().optional(),
+
+  // Cache
+  CACHE_ENABLED: z.coerce.boolean().default(true),
+  CACHE_TTL_SECONDS: z.coerce.number().int().min(10).default(900),
+  REDIS_URL: z.string().url().optional(),
+  REDIS_PASSWORD: z.string().optional(),
+  REDIS_KEY_PREFIX: z.string().default('aidash:cache:'),
+
+  // App PostgreSQL
+  POSTGRES_APP_URL: z.string().default('postgresql://app_user:changeme@localhost:5433/portainer_dashboard'),
+  POSTGRES_APP_MAX_CONNECTIONS: z.coerce.number().int().min(1).max(200).default(20),
+
+  // TimescaleDB (metrics + KPI storage)
+  TIMESCALE_URL: z.string().default('postgresql://metrics_user:changeme@localhost:5432/metrics'),
+  TIMESCALE_MAX_CONNECTIONS: z.coerce.number().int().min(1).max(200).default(50),
+  TIMESCALE_REPORTS_MAX_CONNECTIONS: z.coerce.number().int().min(1).max(50).default(5),
+  METRICS_RAW_RETENTION_DAYS: z.coerce.number().int().min(1).default(7),
+  METRICS_ROLLUP_5MIN_RETENTION_DAYS: z.coerce.number().int().min(1).default(30),
+  METRICS_ROLLUP_1HOUR_RETENTION_DAYS: z.coerce.number().int().min(1).default(90),
+  METRICS_ROLLUP_1DAY_RETENTION_DAYS: z.coerce.number().int().min(1).default(365),
+
+  // Server
+  PORT: z.coerce.number().int().min(1).max(65535).default(3051),
+  LOG_LEVEL: z.enum(['trace', 'debug', 'info', 'warn', 'error', 'fatal']).default('info'),
+  // HTTP/2 (opt-in, requires TLS cert/key)
+  HTTP2_ENABLED: z.coerce.boolean().default(false),
+  TLS_CERT_PATH: z.string().optional(),
+  TLS_KEY_PATH: z.string().optional(),
+
+  // Notifications — Teams
+  TEAMS_WEBHOOK_URL: z.string().url().optional(),
+  TEAMS_NOTIFICATIONS_ENABLED: z.coerce.boolean().default(false),
+
+  // Notifications — Discord
+  DISCORD_WEBHOOK_URL: z.string().url().optional(),
+  DISCORD_NOTIFICATIONS_ENABLED: z.coerce.boolean().default(false),
+
+  // Notifications — Telegram
+  TELEGRAM_BOT_TOKEN: z.string().optional(),
+  TELEGRAM_CHAT_ID: z.string().optional(),
+  TELEGRAM_NOTIFICATIONS_ENABLED: z.coerce.boolean().default(false),
+
+  // Notifications — Email
+  SMTP_HOST: z.string().optional(),
+  SMTP_PORT: z.coerce.number().int().min(1).max(65535).default(587),
+  SMTP_SECURE: z.coerce.boolean().default(true),
+  SMTP_USER: z.string().optional(),
+  SMTP_PASSWORD: z.string().optional(),
+  SMTP_FROM: z.string().default('AI Portainer Dashboard <notifications@example.com>'),
+  EMAIL_NOTIFICATIONS_ENABLED: z.coerce.boolean().default(false),
+  EMAIL_RECIPIENTS: z.string().default(''),
+
+  // Webhooks
+  WEBHOOKS_ENABLED: z.coerce.boolean().default(false),
+  WEBHOOKS_MAX_RETRIES: z.coerce.number().int().min(0).max(10).default(5),
+  WEBHOOKS_RETRY_INTERVAL_SECONDS: z.coerce.number().int().min(10).default(60),
+
+  // Harbor Registry (Vulnerability Management)
+  HARBOR_API_URL: optionalUrl,
+  HARBOR_ROBOT_NAME: z.string().optional(),
+  HARBOR_ROBOT_SECRET: z.string().optional(),
+  HARBOR_VERIFY_SSL: z.string().default('true').transform((v) => v === 'true' || v === '1'),
+  HARBOR_SYNC_ENABLED: z.coerce.boolean().default(false),
+  HARBOR_SYNC_INTERVAL_MINUTES: z.coerce.number().int().min(5).max(1440).default(30),
+  HARBOR_CONCURRENCY: z.coerce.number().int().min(1).max(50).default(5),
+
+  // Image Staleness
+  IMAGE_STALENESS_CHECK_ENABLED: z.coerce.boolean().default(true),
+  IMAGE_STALENESS_CHECK_INTERVAL_HOURS: z.coerce.number().int().min(1).default(24),
+
+  // Prompt Injection Guard
+  LLM_PROMPT_GUARD_STRICT: z.string().default('true').transform((v) => v === 'true' || v === '1'),
+  PROMPT_GUARD_NEAR_MISS_ENABLED: z.string().default('true').transform((v) => v === 'true' || v === '1'),
+  PROMPT_GUARD_NEAR_MISS_LOW_STRICT: z.coerce.number().min(0).max(1).default(0.2),
+  PROMPT_GUARD_NEAR_MISS_HIGH_STRICT: z.coerce.number().min(0).max(1).default(0.4),
+  PROMPT_GUARD_NEAR_MISS_LOW_RELAXED: z.coerce.number().min(0).max(1).default(0.3),
+  PROMPT_GUARD_NEAR_MISS_HIGH_RELAXED: z.coerce.number().min(0).max(1).default(0.5),
+
+  // MCP (Model Context Protocol)
+  MCP_TOOL_TIMEOUT: z.coerce.number().int().min(1).max(600).default(60),
+  LLM_MAX_TOOL_ITERATIONS: z.coerce.number().int().min(1).max(20).default(10),
+
+  // eBPF Trace Ingestion (Grafana Beyla)
+  TRACES_INGESTION_ENABLED: z.string().default('false').transform((v) => v === 'true' || v === '1'),
+  TRACES_INGESTION_API_KEY: z.string().default(''),
+
+  // OpenTelemetry Span Export (OTLP/HTTP JSON)
+  OTEL_EXPORTER_ENABLED: z.string().default('false').transform((v) => v === 'true' || v === '1'),
+  OTEL_EXPORTER_ENDPOINT: z.string().url().optional(),
+  OTEL_EXPORTER_HEADERS: z.string().optional(),
+  OTEL_EXPORTER_BATCH_SIZE: z.coerce.number().int().min(1).max(1000).default(100),
+  OTEL_EXPORTER_FLUSH_INTERVAL_MS: z.coerce.number().int().min(500).max(60000).default(5000),
+
+  // Log Shipping (Elasticsearch)
+  LOG_SHIPPING_ENABLED: z
+    .string()
+    .default('false')
+    .transform((v) => v === 'true' || v === '1'),
+  LOG_SHIPPING_ENDPOINT: z.string().url().optional(),
+  LOG_SHIPPING_INDEX_PREFIX: z.string().default('dashboard-logs'),
+  LOG_SHIPPING_USERNAME: z.string().optional(),
+  LOG_SHIPPING_PASSWORD: z.string().optional(),
+  LOG_SHIPPING_BATCH_SIZE: z.coerce.number().int().min(1).default(100),
+  LOG_SHIPPING_FLUSH_INTERVAL_MS: z.coerce.number().int().min(500).default(5000),
+
+  // Scalability Limits (#544, #547)
+  INSIGHTS_RETENTION_DAYS: z.coerce.number().int().min(1).default(7),
+  MAX_LLM_HISTORY_MESSAGES: z.coerce.number().int().min(1).default(50),
+  LOG_ANALYSIS_CONCURRENCY: z.coerce.number().int().min(1).max(20).default(3),
+
+  // Rate Limiting
+  API_RATE_LIMIT: z.coerce.number().int().min(10).default(
+    process.env.NODE_ENV === 'production' ? 600 : 1200
+  ),
+  LOGIN_RATE_LIMIT: z.coerce.number().int().min(1).default(
+    process.env.NODE_ENV === 'production' ? 5 : 30
+  ),
+});
+
+export type EnvConfig = z.infer<typeof envSchema>;
